it works, a bit more HQ
Yet to do:
revert to plain VFR-CFR dups and drops in ffms !!! per advice from Donald Graft on what we see in real life with phone camera content so far ...
so dups/drops to the motion estimated average framerate ... if that is not reported by medianifo etc, try (A) double the target framerate and go from there.  If that is what the code already does, leave it alone.  We know there's an issue because most phone cameras to 30p or 60p whereas PAl is 25p ...Tempted to just leave it alone because we already do (A)

no cleanup as yet

h,264 4 Mbps vs 5 Mbps @ 1080p25 ... what the heck, leave at 5 Mbps

See what happens with 4k  2160p = 4k	 3840 x 2160 @ 20 Mbps ... to low a bitrate > should be OK with a slideshow though (video clips aside)